auto_build: More hardening

+++ auto_build/rename_build.sh
@@ -6,9 +6,14 @@
 
 if [ "$#" -lt 3 ]; then echo Usage: $0 build version timestamp; exit; fi
 
-in_z=$1 # expecting a tar.bz2 file here
+in_z=$1
 version=$2
 timestamp=$3
+
+if [[ ! $in_z =~ \.tar\.bz2$ ]]; then
+    echo Expecting $in_z to be a .tar.bz2 archive
+    exit 1
+fi
 
 # Go to asp_tarballs and make things relative to that directory
 if [ ! -d "asp_tarballs" ]; then
@@ -37,12 +42,21 @@ fi
 in=${in_z/.tar.bz2/}
 out=${out_z/.tar.bz2/}
 
+# Unpack
 rm -rf $in $out
 if [ -d "$in" ] || [ -d "$out" ]; then exit 1; fi # could not wipe
 bzip2 -dc $in_z | tar xfv - > /dev/null 2>&1
+if [ "$?" -ne 0 ]; then echo "Unpacking failed"; exit 1; fi
 mv $in $out
-cp -f ../dist-add/asp_book.pdf $out # Copy the ASP book
+
+# Copy the ASP book
+doc=../dist-add/asp_book.pdf
+if [ ! -f "$doc" ]; then echo "Could not find $(pwd)/$doc"; exit 1; fi
+cp -f $doc $out 
+
+# Pack back
 tar cf $out_z --use-compress-prog=pbzip2 $out
+if [ "$?" -ne 0 ]; then echo "Packing back failed"; exit 1; fi
 rm -rf $in $out
 
 echo asp_tarballs/$out_z
